requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
psycopg2>=2.9,<3.0
django-cors-headers>=3.10,<4.0
djangorestframework-simplejwt>=5.0,<6.0
drf-yasg>=1.20,<2.0
django-filter>=23.1,<24.0
psycopg2-binary<3.0,>=2.9
